+++ CHANGELOG.md
@@ -3,12 +3,14 @@
 ADDED:
 
 - Identity Map
+- Support for retrieving multiple levels of relationships in the same query
 - Various methods to `SelectQueryBuilder` to reflect changes in Larva v0.3
 
 CHANGED:
 
 - Updated minimum Larva version requirement to v0.3.0
 - Optimized retrieval of relationships
+
 REMOVED:
 
 - `ModelInterface::isAutoIncremented()` method